fix: resolve knip and lint errors
- Remove unused generateEncryptionKey function export
- Remove unused decryptClaudeToken export (function duplicated in get-user-claude-token.ts)
- Ignore get-user-claude-token.ts in knip config (will be used by E2B executor)
- Clean up knip configuration
